Implemented mission choices within the debriefing screen
We now have the option of choosing different missions from within the
debriefing screen.  With the levels.json file we add a fourth argument
to the debriefing which is an array of level numbers that can be
reached from that debriefing scree.  It needs to be either 1 or 2
elements long.  If it's only 1 element long the debriefing looks and
acts as it previously did (pressing enter to continue).  If there are 2
elements the 'enter' button is disabled and instead the user must
select a or be to pick their next mission.

I also added a background image for the briefing and debriefing screens.